+++ laboratory/runtime/src/main/kotlin/io/mehow/laboratory/OptionFactory.kt
@@ -13,6 +13,12 @@ public interface OptionFactory {
   public fun create(key: String, name: String): Feature<*>?
 
   /**
+   * Returns the feature option associated with the provided key and binary value, or `null` if none
+   * match.
+   */
+  public fun create(key: String, binaryValue: Boolean): Feature<*>?
+
+  /**
    * Combines this factory with another. The resulting factory first checks this factory, and then
    * the provided one if no match is found.
    */
@@ -20,6 +26,9 @@ public interface OptionFactory {
     object : OptionFactory {
       override fun create(key: String, name: String) =
         this@OptionFactory.create(key, name) ?: factory.create(key, name)
+
+      override fun create(key: String, binaryValue: Boolean) =
+        this@OptionFactory.create(key, binaryValue) ?: factory.create(key, binaryValue)
     }
 
   public companion object